+++ src-tauri/src/antivirus/scan.rs
@@ -1,13 +1,13 @@
 use specta::specta;
 use std::path::PathBuf;
-use tauri::{command, Emitter};
+use tauri::{Emitter, Manager, command};
 
 use crate::{
     helpers::{
         history::append_scan_history,
         log::{initialize_log_with_id, log_err},
         new_id, resolve_command,
-        scan::{estimate_total_files, get_root_path, run_scan, SCAN_PROCESS},
+        scan::{SCAN_PROCESS, estimate_total_files, get_main_paths, get_root_path, run_scan},
         silent_command,
     },
     types::{
@@ -24,11 +24,12 @@ pub fn get_startup_scan(state: tauri::State<StartupScan>) -> StartupScan {
 
 #[command]
 #[specta(result)]
-pub fn start_main_scan(app: tauri::AppHandle, paths: Vec<PathBuf>) -> Result<(), String> {
+pub fn start_main_scan(app: tauri::AppHandle) -> Result<(), String> {
     let log_id = new_id();
     let log =
         initialize_log_with_id(&app, LogCategory::Scan, &log_id).map_err(|e| e.to_string())?;
     let log_file = log.file;
+    let paths = get_main_paths(app.path());
     append_scan_history(
         &app,
         HistoryItem {