+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   - Enable Scheduler UI
 ### Improved
 - The `useSettings` hook
+- The Scan Start Logic
+### Changes
+- Moved the paths fetch logic for the main scan into the backend
 
 ## [0.3.2] - 2026-02-01 (February Pre-Release)
 ### Added
@@ -38,6 +41,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Moved Real-time Scan Settings into Advanced Settings
 - Renamed Protection Settings into Exclusion Settings
 - Exclusion Settings Icons
+- Changed the Definition update age from 3 days to 7 days (1 week)
 
 ## [0.3.1] - 2026-01-28
 ### Added